@@ -47,7 +47,7 @@
 		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.6",
 		"supersearch": "^0.0.1",
-		"svelte": "^5.1.9",
+		"svelte": "^5.18.0",
 		"svelte-check": "^4.0.4",
 		"tailwindcss": "^3.4.10",
 		"tslib": "^2.6.3",
