@@ -14,7 +14,8 @@
 		"test:unit": "vitest",
 		"lint": "eslint . && prettier --check .",
 		"format": "prettier --write .",
-		"prepare": "npm run package"
+		"prepare": "npm run package",
+		"watch": "svelte-kit sync && svelte-package -w"
 	},
 	"files": [
 		"dist",
@@ -47,16 +48,17 @@
 		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/eslint": "^9.6.0",
 		"codemirror": "^6.0.1",
+		"codemirror-lang-lxlquery": "^0.1.0",
 		"eslint": "^9.7.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.36.0",
-		"esm-env": "^1.1.4",
+		"esm-env": "^1.2.1",
 		"globals": "^15.0.0",
 		"prettier": "^3.3.2",
 		"prettier-plugin-css-order": "^2.1.2",
 		"prettier-plugin-svelte": "^3.2.6",
 		"publint": "^0.2.0",
-		"svelte": "^5.1.9",
+		"svelte": "^5.18.0",
 		"svelte-check": "^4.0.0",
 		"typescript": "^5.0.0",
 		"typescript-eslint": "^8.0.0",
