@@ -22,7 +22,10 @@
 		{ value: `-_sortKeyByLang.${$page.data.locale}`, label: $page.data.t('sort.alphaDesc') },
 		{ value: '-@reverse.instanceOf.publication.year', label: $page.data.t('sort.publicationDesc') },
 		{ value: '@reverse.instanceOf.publication.year', label: $page.data.t('sort.publicationAsc') },
-		{ value: '-reverseLinks.totalItems', label: $page.data.t('sort.linksDesc') }
+		{
+			value: '-reverseLinks.totalItemsByRelation.itemOf.instanceOf',
+			label: $page.data.t('sort.holdingsDesc')
+		}
 	];
 
 	function handleSortChange(e: Event) {
@@ -51,7 +54,7 @@
 	{@const facets = searchResult.facetGroups}
 	{@const numHits = searchResult.totalItems}
 	{@const filterCount = getFiltersCount(searchResult.mapping)}
-	{#if searchResult.predicates.length > 0}
+	{#if searchResult.predicates.length}
 		<nav
 			class="border-b border-primary/16 px-4 md:flex lg:px-6"
 			aria-label={$page.data.t('search.selectedFilters')}
@@ -148,7 +151,7 @@
 					{:else}
 						<span class="hits-count">{$page.data.t('search.noResults')}</span>
 					{/if}
-					{#if searchResult._spell.length > 0}
+					{#if searchResult._spell.length}
 						<span class="suggest">
 							{#each searchResult._spell as suggestion (suggestion.label)}
 								{$page.data.t('search.didYouMean')}
