@@ -45,8 +45,7 @@ export default {
 		'@reverse': 'Relation',
 		'meta.encodingLevel': 'Beskrivningsnivå',
 		'rdf:type': 'Typ',
-		workType: 'Typ',
-		instanceType: 'Format',
+		hasInstanceType: 'Format',
 		genreForm: 'Genre/form',
 		itemHeldBy: 'Bibliotek',
 		bibliography: 'Bibliografi',
@@ -96,16 +95,39 @@ export default {
 		hideDetails: 'Visa mindre',
 		occursAs: 'förekommer som',
 		relatedSearchLabel: 'Sök i resultaten',
-		didYouMean: 'Menade du'
+		didYouMean: 'Menade du',
+		showAllResults: 'Visa alla sökresultat',
+		supersearchStartHeader: 'Bygg och förfina din sökning',
+		addAs: 'Lägg till som'
+	},
+	qualifiers: {
+		contributorKey: 'contributor',
+		contributorLabel: 'Medverkan',
+		contributorPlaceholder: 'Selma Lagerlöf, Astrid Lindgren',
+		titleKey: 'titel',
+		titleLabel: 'Titel',
+		titlePlaceholder: 'Röda rummet, Casablanca',
+		languageKey: 'språk',
+		languageLabel: 'Språk',
+		languagePlaceholder: 'Svenska, Engelska, Arabiska',
+		subjectKey: 'ämne',
+		subjectLabel: 'Ämne',
+		subjectPlaceholder: 'Sommaren, Romerska riket',
+		yearKey: 'år',
+		yearLabel: 'År',
+		yearPlaceholder: '2002, 1987-1994',
+		genreFormKey: 'genreForm',
+		genreFormLabel: 'Genre/form',
+		genreFormPlaceholder: 'Skönlitteratur, Tidsskrift'
 	},
 	sort: {
 		sort: 'Sortera',
 		relevancy: 'Relevans',
 		alphaAsc: 'A-Ö',
 		alphaDesc: 'Ö-A',
-		publicationAsc: 'Utgivningsår (äldst först)',
-		publicationDesc: 'Utgivningsår (nyast först)',
-		linksDesc: 'Mest länkad',
+		publicationAsc: 'Äldst först (utgivningsår)',
+		publicationDesc: 'Nyast först (utgivningsår)',
+		holdingsDesc: 'Antal bibliotek',
 		hitsAsc: 'Minst träffar',
 		hitsDesc: 'Flest träffar',
 		yearAsc: 'Äldst först',
