@@ -31,7 +31,8 @@ export enum Fmt {
 // https://github.com/libris/definitions/blob/develop/source/vocab/base.ttl
 export enum Base {
 	Resource = 'Resource',
-	StructuredValue = 'StructuredValue'
+	StructuredValue = 'StructuredValue',
+	Identity = 'Identity'
 }
 
 // https://github.com/libris/definitions/blob/develop/source/vocab/concepts.ttl
