@@ -5,6 +5,7 @@ import { getSupportedLocale } from '$lib/i18n/locales.js';
 import { type ApiError } from '$lib/types/api.js';
 import type { PartialCollectionView } from '$lib/types/search.js';
 import { asResult } from '$lib/utils/search';
+import { DebugFlags } from '$lib/types/userSettings';
 
 export const load = async ({ params, url, locals, fetch }) => {
 	const displayUtil = locals.display;
@@ -16,8 +17,10 @@ export const load = async ({ params, url, locals, fetch }) => {
 		redirect(303, `/`); // redirect to home page if no search params are given
 	}
 
+	const debug = locals.userSettings?.debug?.includes(DebugFlags.ES_SCORE) ? '&_debug=esScore' : '';
+
 	const searchParams = new URLSearchParams(url.searchParams.toString());
-	const recordsRes = await fetch(`${env.API_URL}/find.jsonld?${searchParams.toString()}`, {
+	const recordsRes = await fetch(`${env.API_URL}/find.jsonld?${searchParams.toString()}${debug}`, {
 		// intercept 3xx redirects to sync back the correct _i/_q combination provided by api
 		redirect: 'manual'
 	});
@@ -40,14 +43,7 @@ export const load = async ({ params, url, locals, fetch }) => {
 
 	const result = (await recordsRes.json()) as PartialCollectionView;
 
-	const searchResult = await asResult(
-		result,
-		displayUtil,
-		vocabUtil,
-		locale,
-		env.AUXD_SECRET,
-		url.pathname
-	);
+	const searchResult = await asResult(result, displayUtil, vocabUtil, locale, env.AUXD_SECRET);
 
 	return { searchResult, userSettings };
 };
