@@ -35,6 +35,6 @@ test('url is populated with correct searchparams', async ({ page }) => {
 	await page.getByTestId('main-search').fill('somephrase');
 	await page.getByTestId('main-search').press('Enter');
 	await expect(page).toHaveURL(
-		/_q=somephrase&_limit=20&_offset=0&_sort=&_spell=true&_i=somephrase/
+		/_q=somephrase&_i=somephrase&_limit=20&_offset=0&_sort=&_spell=true/
 	);
 });
