@@ -4,7 +4,7 @@ test('decorated data label visibilty is correct after page navigations', async (
 	// TODO: We should probably mock the required requests but something similar to https://github.com/markjaquith/sveltekit-playwright-fetch-mock would be needed to mock the server-side fetches.
 	await page.goto('/h08ndxddfg5v2pjf');
 	await expect(page.getByText('Medverkan och funktion')).toBeHidden();
-	await page.getByText('Khemiri, Jonas Hassen, 1978-').first().click();
+	await page.getByText('Jonas Hassen Khemiri, 1978-').first().click();
 	await expect(page.getByText('Språk')).toBeVisible();
 });
 
