@@ -1,3 +1,4 @@
 import SuperSearch from '$lib/components/SuperSearch.svelte';
+import lxlQualifierPlugin from '$lib/extensions/lxlQualifierPlugin/index.js';
 
-export { SuperSearch };
+export { SuperSearch, lxlQualifierPlugin };
